[win32] Fix compile errors in the ANSI build.
Fix several Windows-specific format strings for types that are
technically `long`. (These errors didn't show up in the Unicode build.)

In file included from /home/david/programming/rom-properties/src/librpsecure/os-secure_win32.c:18:
/home/david/programming/rom-properties/src/librpsecure/os-secure_win32.c: In function ‘rp_secure_reduce_integrity’:
/home/david/programming/rom-properties/src/librpsecure/os-secure_win32.c:60:38: error: format ‘%u’ expects argument of type ‘unsigned int’, but argument 3 has type ‘DWORD’ {aka ‘long unsigned int’} [-Werror=format=]
   60 |                 _ftprintf(stderr, _T("*** DEBUG: SetProcessIntegrityLevel() failed: %u\n"), dwRet);
      |                                      ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/home/david/programming/rom-properties/src/librpsecure/os-secure_win32.c:60:86: note: format string is defined here
   60 |                 _ftprintf(stderr, _T("*** DEBUG: SetProcessIntegrityLevel() failed: %u\n"), dwRet);
      |                                                                                     ~^
      |                                                                                      |
      |                                                                                      unsigned int
      |                                                                                     %lu

[librpfile] FileSystem_win32.cpp: utf8_to_ansi() isn't available here.

Use MiniU82T's U82T_s(). Note that it doesn't properly convert to ANSI
right now; it just passes the UTF-8 string as-is.

MiniU82T.cpp: U82W_c() should use wstring, not tstring.

AchWin32.cpp: Fix an incorrect use of U82W_s(). Should've been U82T_s().

KeyManagerTab.cpp: s/rp_swprintf()/rp_stprintf()/

FIXME: GUID_fn.c has ANSI implementations of some functions, but it's
not currently linked in, and I don't want to export these symbols...

+++ src/librpbase/img/APNG_dlopen.c
@@ -89,10 +89,10 @@ static int init_apng(void)
 	// Otherwise, this will fail.
 #ifndef NDEBUG
 	_sntprintf(png_dll_filename, _countof(png_dll_filename),
-		_T("libpng%ud.dll"), PNG_LIBPNG_VER_DLLNUM);
+		_T("libpng%ud.dll"), (unsigned int)PNG_LIBPNG_VER_DLLNUM);
 #else
 	_sntprintf(png_dll_filename, _countof(png_dll_filename),
-		_T("libpng%u.dll"), PNG_LIBPNG_VER_DLLNUM);
+		_T("libpng%u.dll"), (unsigned int)PNG_LIBPNG_VER_DLLNUM);
 #endif
 	png_dll_filename[_countof(png_dll_filename)-1] = _T('\0');
 	bRet = GetModuleHandleEx(0, png_dll_filename, &libpng_dll);
